Add dateformat and fix the tests
