+++ incoming-message-handler/package.json
@@ -6,6 +6,7 @@
   "main": "index.js",
   "devDependencies": {
     "@types/aws-lambda": "^8.10.64",
+    "@types/dateformat": "^3.0.1",
     "@types/jest": "^26.0.20",
     "@types/node": "^14.14.28",
     "@types/uuid": "^8.3.0",
@@ -31,6 +32,7 @@
     "@handlers/common": "file:../handlers-common",
     "aws-sdk": "^2.804.0",
     "axios": "^0.21.1",
+    "dateformat": "^4.5.1",
     "https": "^1.0.0",
     "uuid": "^8.3.2",
     "xml2js": "^0.4.23"
