+++ incoming-message-handler/src/use-cases/UploadMessageUseCase.test.ts
@@ -1,7 +1,6 @@
 import { isError } from "@handlers/common"
 import S3Gateway from "../gateways/S3Gateway"
 import { S3Config } from "../types"
-import { getFileName } from "../utils/file"
 import UploadMessageUseCase from "./UploadMessageUseCase"
 
 const config: S3Config = {
@@ -9,16 +8,16 @@ const config: S3Config = {
   S3_REGION: "us-east-1"
 }
 
-const messageId = "message123"
+const messageId = "123456"
 const receivedDate = new Date(2021, 3, 9, 12, 15, 18)
 const gateway = new S3Gateway(config)
 const useCase = new UploadMessageUseCase(gateway)
+const expectedResult = "2021/04/09/12/15/123456.xml"
 
 describe("UploadMessageUseCase", () => {
   it("should save the message", async () => {
     jest.spyOn(gateway, "upload").mockResolvedValue(undefined)
 
-    const expectedResult = getFileName(receivedDate, messageId)
     const result = await useCase.save({ messageId, caseId: "case123", receivedDate, rawXml: "message" })
 
     expect(isError(result)).toBe(false)
@@ -26,7 +25,6 @@ describe("UploadMessageUseCase", () => {
   })
 
   it("should fail when the error is unknown", async () => {
-    const expectedFileName = getFileName(receivedDate, messageId)
     jest.spyOn(gateway, "upload").mockResolvedValue(new Error("An unknown error"))
 
     const result = await useCase.save({ messageId, caseId: "case123", receivedDate, rawXml: "message" })
@@ -34,6 +32,7 @@ describe("UploadMessageUseCase", () => {
     expect(isError(result)).toBe(true)
 
     const error = <Error>result
-    expect(error.message).toBe(`The file ${expectedFileName} could not be saved`)
+
+    expect(error.message).toBe(`The file ${expectedResult} could not be saved`)
   })
 })
