--- incoming-message-handler/src/utils/file.test.ts
@@ -1,13 +1,0 @@
-import { getFileName } from "./file"
-
-describe("file", () => {
-  describe("getFileName", () => {
-    it("should get the full path for the file to be saved", async () => {
-      const date = new Date(2021, 3, 9, 12, 6)
-      const messageId = "123456"
-      const expectedResult = `2021/03/09/12/06/${messageId}.xml`
-      const actualResult = getFileName(date, messageId)
-      expect(actualResult).toBe(expectedResult)
-    })
-  })
-})
